+++ DS3231Plus.h
@@ -16,8 +16,8 @@ DS3231Plus(DS3231Plus&)=delete;
 DS3231Plus(DS3231Plus&&)=delete; 
 
 // Delay and call functions
-void DelayRTCLock(const int& Miliseconds,void (*func_todo)());
-void DelayRTCLock(const int& Miliseconds);
+void DelayRTCLock(const long int& Miliseconds,void (*func_todo)());
+void DelayRTCLock(const long int& Miliseconds);
 
 // Get functions
 String GetTimeString(bool ShowDate=true,bool ShowDow=true,bool ShowTime=true);
